Very preliminary Garrote while Sepsis Buff up Fix
On the main branch, the Sepsis Buff gets completly ignored. With this, it will suggest to Garrote either when the Sepsis Debuff has 3 or less seconds, or if an garrote is in the pandemic window while Sepsis Debuff is up. 
It is preliminary because it does not always work correctly. Sometimes it will suggest Garrote too late, as in, the Sepsis Buff only last 2 or even 1 second. However, the pandemic window extend works well. I just wanted to save it online so I can work around a bit more. 
If you got any suggestions as to why it does not always work correctly, please let me know.