Probably Fixed Garrote Suggestion if SepsisBuff up
In this update, I've adjusted the script to consistently suggest the use of Garrote when there are 3 or less seconds remaining on the Sepsis buff. Although this may initially seem counterintuitive, since one might aim for the full duration of the last Sepsis-enhanced Garrote, especially during the opener with Deathmark, this adjustment in fact minimizes potential disruptions caused by energy shortages that could interfere with a successful Garrote cast. 

For instance, if Envenom is used at the 3-second mark of the Sepsis buff, it could theoretically drain my energy to such an extent that the global cooldown (GCD) extends to 2 seconds. This could cause me to miss the opportunity to cast Garrote.

With this modification, I prioritize consistency and reliability of action recommendations, adopting a 'better safe than sorry' approach.

Looking ahead, it may be possible to implement an energy check in the future, circumventing this whole issue entirely. But for now, I'm satisfied with the improvement this change brings to the script."

Let me know if there's anything else you'd like me to add or adjust!

(Also changed the Shiv suggestion a bit, now always lining with Sepsis, if worth)