Garrote reapply check when Sepsis up soon
This commit introduces a smarter interaction between the Garrote and Sepsis abilities. Now, if Garrote is about to wear off within the next 5 seconds but Sepsis will become available in that same timeframe, the rotation will hold off on suggesting to reapply Garrote. This allows Sepsis to be used more optimally.

Additionally, the requirement for using the Sepsis buff has been reduced to 2.5 seconds. This change facilitates better timing and efficiency in using the buff, leading to an overall smoother and more effective rotation.