Garrote after Sepsis Temp Fix + Shiv in Opener
Added a global variable _G.MyAddon_ShivCast to track the casting status of the ability 'Shiv'. This variable is initialized as false at the start of the script.
Modified the condition for casting 'Shiv' to include a check for the 'Sepsis' debuff on the target and the _G.MyAddon_ShivCast variable. This ensures 'Shiv' is only suggested if the target does not have the 'Sepsis' debuff for more than seven seconds or if 'Shiv' has not been cast yet.
Updated _G.MyAddon_ShivCast to true after 'Shiv' is cast, indicating that 'Shiv' has been cast and preventing further suggestions to cast 'Shiv' while the 'Sepsis' debuff is active on the target for more than seven seconds.
These changes optimize the usage of 'Shiv' in the rotation by preventing its suggestion when the 'Sepsis' debuff is active on the target for more than seven seconds, except for the first 'Shiv' cast.
This is only a workaround. Will try to force it to use garrote after casting sepsis. WiP