Updated SimC Lines and simmed changes
All my changes result in about a 570 dps increase.
200 from the symbol change
200 from the SnD change,
100 from the vanish change
70 from Secret at 3.1 seconds change

I will sim some different SnD numbers to check if there is an improvment later
Oh and I simmed my check for no dance in the rupture related code - turns out: nothing changes in terms of damage, meaning SkipRupture just does not work correctly in game although it is correctly implemented. My Shadowdance check just enforces this behaviour in hero rotation. There are scenarios in the sim where it would use Rupture in Dance, but these are rare - the same goes for SnD in Dance, which is also rare. I am unsure as to how this is relevant as I sometimes see a suggestion for Dance into SnD and after pressing dance, it shows GB, but I am too tired to make any judgments. If I had to guess, I would say it has something to do with pooling finishers or the macro table, which I dont know what that even is lol. Anyways, seems to be good and working.