PTR Testing Files
Based on my initial tests, there are no Lua errors. However, the rotation is not yet optimal. Lag on the PTR makes precise evaluation challenging, but the rotation generally appears to be accurate, with minor exceptions. I've only tested with one specific build so far (BUQAt5bPuh6pO3KIPROgVVbp4CAEEAAAAAASLJJlEQKcgkIikEJJRi0SJRiWSCSgIpkSAAAAIBB).
There are two notable issues:

Pre-combat activation of Symbols of Death is not functioning.
The last global cooldown in Shadow Dance suggests a Shadowstrike, which gets cancelled due to insufficient time.
If you are familiar with the PTR rotation, I welcome your insights for further refinement. I will continue to review the rotation for any discrepancies and further updates. Assistance in this regard would be highly appreciated.

To install the updated Lua files:

Navigate to the PTR Addons folder and locate the HeroRotation sub-folders.
Find the HeroRotationRogue directory.
Download RoguePTR.lua and SubtletyPTR.lua.
Insert the downloaded files in the folder
Delete the original files, i.e Rogue.lua and Subtlety.lua.
Rename RoguePTR.lua to Rogue.lua and SubtletyPTR.lua to Subtlety.lua.
Log into the game to apply the changes.