Dont use; some issues
There appear to be discrepancies between the simulation results and the current Lua implementation. For instance, the simulation suggests using Vanish once in the initial minute, while the Lua code seems to prompt for both charges in close succession. Furthermore, the Symbols of Death (SoD) implementation appears to be malfunctioning; it is entirely absent in the opener and only sporadically functional thereafter. I would greatly appreciate any insights or contributions to address these issues. Please feel free to comment under the commit. Thank you.