10.2 | PLEASE DOWNLOAD SETTINGS.LUA AS WELL
Updated to the newest settings, also included a OffGCD check for Flag.
Took over the offical Assa lua. Will work on it in the coming days. Does seem to be working, however some things are off.
Sub is updated to work with the newest Settings.lua. 
Rogue.lua updated to reflect changes.