-- read description -- "Bugfix" for Vanish and Vanish in Dance Option (pls download settings.lua as well) -- read description --
Now prevents Vanish to trigger too soon when Flag was used but Secret had too much cd, resulting in vanishing way before even being in dance. (may happen at 1:00-1:10)
Added the option to allow Vanish in Dance. Be aware tho: in sims, its about an 2k damage loss. I added it because I saw that the majority of Top-Logs do it on bosses such as Igira or Nymue - my guess is because it allows better uses? Anyway, this setting shouldnt change any of the custom conditions for specific bosses. Oh and another thing - a solo flag, at least in sims, is ALWAYS worse than the "Solo"-Flag with Vanish - idk why, but it is, apparently.
The very first Vanish, so when you have 2 stacks of vanish, will be the exact same as before. Afterward, it will show Vanish in Dance to pull Blades - or, to be a bit more precise - to reduce the CD of Flag since only Flag is actually checked in the condtion. 
I am aware that Vanish sometimes triggers late but still in Dance, so not directly after - when I tested the condition in the sims, it, in fact, did the very same thing and I cant quite pinpoint as to why. So, maybe someone else has an idea.

Very important:

IF YOU HAVE ANY ISSUES, REACH OUT TO ME ON DISCORD DM. THE OFFICIAL DISCORD WILL NOT BE ABLE TO HELP.