Compatibility update to the official verision. Read descrption
You NEED to update the commons.lua in the hero rotation folder. If you do not know how that works, update the addon and then redownload the Lua files you want. Rememeber - you will NEED Settings.lua and Rogue.lua if you want to play any of the lua specs from my fork. If it says 1.6 when loading the game, you are on the correct version.