Exclude Dispatch From Stealth-Windows If BtE Is Castable
Sometimes, the addon would suggest Dispatch near the end of a Stealth-Window although is BtE ready. With this change, this should not happen anymore. 
Often times, you would still be able to hit the BtE within the window. I cant quite pinpoint as to why that happens, but it sometimes did.