Fix for Dispatch in Stealth
Now will not show Dispatch in stealth windows unless you do not have BtE ready. It seems to be a pooling issue.
There is one more "bug" in stealth windows - Roll the Bones will be shown a bit too early, showing on the last GCD in a stealth window instead of BtE when the KiR condition is fullfilled. No idea how to fix that one yet.