Experimental Update...
testing a new setting with which you can limit the amount of Garrotes and Ruptures spread during stealth windows. It shouldnt change too much, besides allowing more CT/Env to be used in Stealth + more use of FoK and Ambush during Stealth. Garrote is still a bit spammy, but Ill work on it.